+++ agent/killer.go
@@ -1,4 +1,4 @@
-package app
+package agent
 
 import (
 	"context"
