+++ cmd/client/run.go
@@ -6,8 +6,9 @@ import (
 	"os/signal"
 	"syscall"
 
-	"github.com/holdno/gopherCron/app"
+	"github.com/holdno/gopherCron/agent"
 	"github.com/holdno/gopherCron/config"
+	"github.com/holdno/gopherCron/pkg/warning"
 	"github.com/holdno/gopherCron/utils"
 )
 
@@ -19,15 +20,15 @@ func initConf(filePath string) *config.ServiceConfig {
 
 func Run(opts *SetupOptions) error {
 	// 加载配置
-	client := app.NewClient(opts.ConfigPath)
+	client := agent.NewClient(opts.ConfigPath)
 
 	restart := func() {
 		defer func() {
 			if r := recover(); r != nil {
 				ip, _ := utils.GetLocalIP()
-				_ = client.Warning(app.WarningData{
+				_ = client.Warning(warning.WarningData{
 					Data:    fmt.Sprintf("agent %s down", ip),
-					Type:    app.WarningTypeSystem,
+					Type:    warning.WarningTypeSystem,
 					AgentIP: client.GetIP(),
 				})
 			}
@@ -49,7 +50,7 @@ func Run(opts *SetupOptions) error {
 	return nil
 }
 
-func waitingShutdown(c app.Client) {
+func waitingShutdown(c agent.Client) {
 	stopSignalChan := make(chan os.Signal, 1)
 	signal.Notify(stopSignalChan, os.Interrupt, os.Kill, syscall.SIGTERM, syscall.SIGINT)
